Move optimizer failure test
